+++ app/controllers/concerns/crappy.rb
@@ -1,0 +1,96 @@
+require 'watir' # Crawler
+require 'pry' # Ruby REPL
+require 'selenium-webdriver'
+require 'chromedriver-helper'
+require_relative 'credentials' # Pulls in login credentials from credentials.rb
+
+username = iusername
+password = password
+users = ["nfl", "baseball", "nba", "flowers"]
+follow_counter = 0
+unfollow_counter = 0
+MAX_UNFOLLOWS = 200
+start_time = Time.now
+
+# Open Browser, Navigate to Login page
+browser = Watir::Browser.new :chrome, switches: ['--incognito']
+browser.goto "instagram.com/accounts/login/"
+
+# Navigate to Username and Password fields, inject info
+puts "Logging in..."
+browser.text_field(:name => "username").set ""
+browser.text_field(:name => "password").set ""
+
+# Click Login Button
+browser.button(:class => ['sqdOP', 'L3NKy', 'y3zKF']).click
+browser.button(:class => ['aOOlW', 'HoLwm']).click #not now popup
+sleep(2)
+puts "We're in #hackerman"
+
+# Continuous loop to run until you've unfollowed the max people for the day
+loop do
+  users.each { |val|
+    # Navigate to user's page
+    browser.goto "instagram.com//"
+
+    # If not following then follow
+    if browser.button(:class => ['_5f5mN','jIbKX','_6VtSN','yZn4P']).exists?
+      ap "Following ilike4u326"
+      browser.button(:class => ['_5f5mN','jIbKX','_6VtSN','yZn4P']).click
+      follow_counter += 1
+    elsif browser.button(:class => ['_5f5mN', '-fzfL','_6VtSN', 'yZn4P']).exists?
+      ap "Unfollowing #{val}"
+      browser.button(:class => ['_5f5mN', '-fzfL','_6VtSN', 'yZn4P']).click
+      browser.button(:class => ['aOOlW','-Cab_']).click
+      unfollow_counter += 1
+    end
+    sleep(1.0/2.0) # Sleep half a second to not get tripped up when un/following many users at once
+  }
+  puts "--------- #{Time.now} ----------"
+  break if unfollow_counter >= MAX_UNFOLLOWS
+  sleep(30) # Sleep 1 hour (3600 seconds)
+end
+
+ap "Followed #{follow_counter} users and unfollowed #{unfollow_counter} in #{((Time.now - start_time)/60).round(2)} minutes"
+
+# Leave this in to use the REPL at end of program
+# Otherwise, take it out and program will just end
+Pry.start(binding)
+
+# Top 100 users on Instagram
+# users = ['instagram', 'selenagomez', 'arianagrande', 'taylorswift', 'beyonce', 'kimkardashian', 'cristiano', 'kyliejenner', 'justinbieber', 'kendalljenner', 'nickiminaj', 'natgeo', 'neymarjr', 'nike', 'leomessi','khloekardashian', 'mileycyrus', 'katyperry', 'jlo', 'ddlovato', 'kourtneykardash', 'victoriasecret', 'badgalriri', 'fcbarcelona', 'realmadrid', 'theellenshow', 'justintimberlake', 'zendaya' 'caradelevingne', '9gag', 'chrisbrownofficial', 'vindiesel', 'champagnepapi', 'davidbeckham', 'shakira', 'gigihadid', 'emmawatson', 'jamesrodiguez10', 'kingjames', 'garethbale11', 'nikefootball', 'adele', 'zacefron', 'vanessahudgens', 'ladygaga', 'maluma', 'nba', 'nasa', 'rondaldinho', 'luissuarez9', 'zayn', 'shawnmendes', 'adidasfootball', 'brumarquezine', 'hm', 'harrystyles','chanelofficial', 'ayutingting92', 'letthelordbewithyou', 'niallhoran', 'anitta', 'hudabeauty', 'camerondallas', 'adidasoriginals', 'marinaruybarbosa', 'lucyhale', 'karimbenzema', 'princessyahrini', 'zara', 'nickyjampr', 'onedirection', 'andresiniesta8', 'raffinagita1717', 'krisjenner', 'manchesterunited', 'natgeotravel', 'marcelottwelve', 'deepikapadukone', 'snoopdogg', 'davidluiz_4', 'kalbiminrozeti', 'priyankachopra', 'ashleybenson', 'shaym', 'lelepons', 'prillylatuconsina96','louisvuitton','britneyspears', 'sr4official', 'jbalvin', 'laudyacynthiabella', 'ciara', 'stephencurry30', 'instagrambrasil']
+
+photo_target  = browser.div(:class => ['_9AhH0']).click
+like = browser.span(:class => ['fr66n']).click
+unlike = browser.span(:class => ['FY9nT', 'fr66n']).click
+search = browser.text_field(:value => "").set "test"
+validate hashtag = browser.send_keys :enter
+
+
+def login
+  browser = Watir::Browser.new :chrome, switches: ['--incognito']
+  sleep(1)
+  browser.goto "instagram.com/accounts/login/"
+  puts "Logging in..."
+  browser.text_field(:name => "username").set ""
+  browser.text_field(:name => "password").set ""
+  sleep(1)
+  browser.button(:class => ['sqdOP', 'L3NKy', 'y3zKF']).click
+  sleep(1)
+  if browser.button(:class => ['aOOlW', 'HoLwm']).exists?
+    browser.button(:class => ['aOOlW', 'HoLwm']).click
+  sleep(1)
+end
+
+def liking
+  browser.goto "instagram.com/karliekloss/"
+  sleep(3.03)
+  browser.div(:class => ['_9AhH0']).click
+  sleep(3.09)
+  browser.span(:class => ['fr66n']).click
+  sleep(3.12)
+end
+
+def following
+
+end
